requirements.txt.txt
python-telegram-bot
openai
google-generativeai
requests
psycopg2-binary